+++ src/main/java/springJpaBoard/Board/api/apicontroller/MemberApiController.java
@@ -18,23 +18,19 @@ import org.springframework.web.bind.annotation.*;
 import springJpaBoard.Board.Error.Message;
 import springJpaBoard.Board.Error.StatusEnum;
 import springJpaBoard.Board.Error.exception.UserException;
-import springJpaBoard.Board.SessionConst;
 import springJpaBoard.Board.domain.Member;
 import springJpaBoard.Board.domain.argumenresolver.Login;
 import springJpaBoard.Board.repository.search.MemberSearch;
 import springJpaBoard.Board.service.BoardService;
 import springJpaBoard.Board.service.MemberService;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-import javax.validation.Valid;
 import java.nio.charset.Charset;
 import java.util.List;
 
 import static java.util.stream.Collectors.toList;
 import static springJpaBoard.Board.controller.boarddto.BoardDto.MyPostsResponse;
-import static springJpaBoard.Board.controller.memberdto.AuthDto.LoginRequest;
-import static springJpaBoard.Board.controller.memberdto.MemberDto.*;
+import static springJpaBoard.Board.controller.memberdto.MemberDto.MemberResponse;
+import static springJpaBoard.Board.controller.memberdto.MemberDto.ModifyMemberRequest;
 
 @RestController
 @RequiredArgsConstructor
@@ -44,72 +40,6 @@ public class MemberApiController {
 
     private final MemberService memberService;
     private final BoardService boardService;
-
-    /* 회원가입 */
-    @PostMapping
-    public ResponseEntity join(@RequestBody @Validated CreateMemberRequest memberRequestDTO, BindingResult result) {
-        if (result.hasErrors()) {
-            throw new UserException("회원가입: 회원 정보 입력 오류");
-        }
-
-        Long id = memberService.join(memberRequestDTO);
-
-        Message message = new Message(StatusEnum.CREATED, "회원 가입 성공", id);
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(new MediaType("application", "json", Charset.forName("UTF-8")));
-
-
-        return new ResponseEntity<>(message, headers, HttpStatus.CREATED);
-    }
-
-    /* 회원 로그인 */
-    // TODO - redirectURL 해결
-    // 현재 생각한 방법: Result 타입에 redirectURL을 추가해서 같이 반환
-    @PostMapping("/login")
-    public ResponseEntity login(@RequestBody @Valid LoginRequest loginRequest, BindingResult result,
-                                @RequestParam(defaultValue = "/") String redirectURL, HttpServletRequest request) {
-        if (result.hasErrors()) {
-            throw new UserException("로그인: 아이디 또는 비밀번호 오류");
-        }
-
-        Member loginMember = memberService.login(loginRequest.loginId(), loginRequest.password());
-
-        if (loginMember == null) {
-            result.reject("loginFail", "아이디 또는 비밀번호가 맞지 않습니다.");
-            throw new UserException("로그인: 아이디 또는 비밀번호 오류");
-        }
-
-
-        /*세션이 있으면 있는 세션 반환, 없으면 신규 세션 생성*/
-        HttpSession session = request.getSession();
-        /*세션에 로그인 회원 정보 보관*/
-        session.setAttribute(SessionConst.LOGIN_MEMBER, loginMember);
-
-        Message message = new Message(StatusEnum.OK, "로그인 성공", loginMember.getLoginId());
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(new MediaType("application", "json", Charset.forName("UTF-8")));
-
-        return new ResponseEntity<>(message, headers, HttpStatus.OK);
-    }
-
-    /* 회원 로그아웃 */
-    @PostMapping("/logout")
-    public ResponseEntity<Message> logout(HttpServletRequest request) {
-        /*세션 삭제*/
-        HttpSession session = request.getSession(false);
-
-        if (session != null) {
-            session.invalidate(); //세션 강제 종료
-
-            Message message = new Message(StatusEnum.OK, "회원 로그아웃 성공");
-            HttpHeaders headers = new HttpHeaders();
-            headers.setContentType(new MediaType("application", "json", Charset.forName("UTF-8")));
-
-            return new ResponseEntity<>(message, headers, HttpStatus.OK);
-        }
-
-        throw new IllegalStateException("세션이 존재하지 않습니다.");
-    }
 
     /* 회원 목록 */
     @GetMapping
@@ -129,8 +59,7 @@ public class MemberApiController {
     }
 
 
-    /* 회원 수정 */
-    @GetMapping("/edit/{memberId}")
+    /* 회원 수정 */@GetMapping("/edit/{memberId}")
     public ResponseEntity updateForm(@PathVariable Long memberId, @Login Member loginMember) {
 
         Member member = memberService.findOne(memberId);
